test(core): address PR feedback and expand tag management tests
- Move all tests into tag.test.ts (delete tag_management.test.ts)
- Use ctx.assert.list_partial for alias/parent/children assertions
- Fix alias_create/parent_create to handle missing source tags
  (TagAliasResponse.alias and TagParentResponse.child are now nullable)
- Add test: alias migration when source+target share a media reference
- Add test: multi-level circular parent detection (A->B->C, then C->A)
- Add test: auto cleanup interaction with aliases
- Add test: alias_create when source tag doesn't exist yet
- Add test: parent_create when source tag doesn't exist yet
